+++ tests/testthat/test-utils.R
@@ -1,5 +1,17 @@
 context("Utils")
 
+test_that("testing sf_download_attachment", {
+  queried_attachments <- sf_query("SELECT Body, Name
+                                   FROM Attachment
+                                   WHERE ParentId = '0016A0000035mJ5'")
+  res <- mapply(sf_download_attachment, 
+                queried_attachments$Body, 
+                queried_attachments$Name)
+  expect_true(all(res))
+  # clean up
+  sapply(queried_attachments$Name, file.remove)
+})
+
 test_that("testing input_data validation", {
   input_data <- 1:3
   res1a <- sf_input_data_validation(input_data)